allow floating point numbers for power and steering in motor and move classes

+++ src/le2lejosev3/pblocks/IMotorRotation.java
@@ -37,6 +37,6 @@ public interface IMotorRotation {
 	 * 
 	 * @return the current power level.
 	 */
-	public int measureCurrentPower();
+	public float measureCurrentPower();
 
 }
